Update index.js
In case you'd be interested, I've modified your module's index.js file a bit to add in the 5eDnD calculations for ammunition attack (line 1015-1029) and damage (line 1143 -  1156) bonuses when the ammo is associated with the weapon used to attack.  Could still use some cleaning up and polishing, but it functionality gets the job done for my needs.